Issue #4267: Fixed broken test
