+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/blocks/emptyblock/InputEmptyBlockDefault.java
@@ -1,87 +1,89 @@
-////////////////////////////////////////////////////////////////////////////////
-// Test case file for checkstyle.
-// Created: 2001
-////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks.blocks.emptyblock;
-
-import java.io.*; // star import for instantiation tests
-import java.awt.Dimension; // explicit import for instantiation tests
-import java.awt.Color;
-
-/**
- * Test case for detecting empty block statements.
- * @author Lars Kühne
- **/
-class InputEmptyBlockDefault
-{
-    static {
-        Boolean x = new Boolean(true);
-    }
-
-    {
-        Boolean x = new Boolean(true);
-        Boolean[] y = new Boolean[]{Boolean.TRUE, Boolean.FALSE};
-    }
-
-    Boolean getBoolean()
-    {
-        return new java.lang.Boolean(true);
-    }
-
-    void exHandlerTest()
-    {
-        try {
+public class InputEmptyBlockDefault {
+    void method1(int a) {
+        switch (a) {}
+        switch (a) {default: ; } // no violation
+        switch (a) {default: {}} // violation
+        switch (a) {
+            default:
         }
-        finally {
+        switch (a) {
+            default:             // violation
+            {}
         }
-        try {
-        // something
-        }
-        finally {
-            // something
-        }
-        try {
-            ; // something
-        }
-        finally {
-            ; // statement
-        }
-    }
-
-    /** test **/
-    private static final long IGNORE = 666l + 666L;
-
-    public class EqualsVsHashCode1
-    {
-        public boolean equals(int a)
-        {
-            return a == 1;
-        }
-    }
-
-    // empty instance initializer
-    {
-    }
-
-    private class InputBraces {
-        
-    }
-
-    synchronized void foo() {
-        synchronized (this) {} // not OK
-        synchronized (Class.class) { // OK
-            synchronized (new Object()) {
-                // not OK if checking statements
+        switch (a) {
+            default:             // violation if checking statements
+            { // text
             }
         }
     }
-    
-    
-    static {
-       
-    int a = 0;}
-    
-    static {
-        
+
+    void method2(int a) {
+        switch (a) {
+            case 1:a++;
+            case 2:a++;
+            default:             // no violation
+                switch (a) {
+                    default: {   // violation if checking for statements
+
+                    }
+                }
+        }
+    }
+
+    void method3(int a, int b) {
+        switch (a) {
+            case 1: break;
+            default: {} method2(a);  // violation
+        }
+
+        switch (b) {
+            case 2: break;
+            default: method2(b); {}  // no violation
+        }
+
+        switch (a+b) {case 1: break; default: {} ; } // violation
+    }
+
+    void method4(int a, int b) {
+        switch (a) {
+            case 1:
+            default: {}    // violation
+        }
+
+        switch (b) {
+            case 1:
+            default:       // no violation
+        }
+
+        switch (a+b) {
+            default:       // no violation
+            case 1: { }
+        }
+
+        switch (a-b) {
+            case 1:
+            default: {     // violation if checking statements
+
+            } ;
+            case 2: { }
+        }
+    }
+
+    void method5(int a, int b) {
+        switch (a) {
+            case 1:
+            case 2:
+            case 3:
+            default:       // violation
+            {
+            }
+        }
+
+        switch (b) {
+            default:       // no violation
+            case 1:
+            case 2: { } method2(b);
+            case 3:
+        }
     }
 }
